extract relation list count logic into separate file
